move exception to exceptions.py
