+++ zntrack/utils/exceptions.py
@@ -1,0 +1,16 @@
+"""
+This program and the accompanying materials are made available under the terms of the
+Eclipse Public License v2.0 which accompanies this distribution, and is available at
+https://www.eclipse.org/legal/epl-v20.html
+SPDX-License-Identifier: EPL-2.0
+
+Copyright Contributors to the Zincware Project.
+
+Description: ZnTrack custom exceptions
+"""
+
+
+class DVCProcessError(Exception):
+    """DVC specific message for CalledProcessError"""
+
+    pass
